Rename Entities directory to Scene
